coords type changed to float

+++ graph/vertex.py
@@ -1,6 +1,6 @@
 class Vertex:
     """Класс вершиный содержит её имя и координаты"""
-    def __init__(self, name: str, x: int, y: int):
+    def __init__(self, name: str, x: float, y: float):
         self.y = y
         self.x = x
         self.name = name
